% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{arrange.harp_list}
\alias{arrange.harp_list}
\title{Arrange rows by column values}
\usage{
\method{arrange}{harp_list}(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Variables, or
functions of variables. Use \code{\link[dplyr:desc]{desc()}} to sort a variable in descending
order.}

\item{.by_group}{If \code{TRUE}, will sort first by grouping variable. Applies to
grouped data frames only.}
}
\description{
This is the \code{harp_list} method for \code{\link[dplyr]{arrange}()}.
It works in exactly the same way except that a \code{harp_list} is returned.
To use this function, the \code{dplyr} package must be attached and the
\code{.harp_list} suffix can be dropped.
}
