% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\name{as_det}
\alias{as_det}
\title{Convert an ensemble forecast to deterministic.}
\usage{
as_det(x, member = NULL, sub_model = NULL)
}
\arguments{
\item{x}{A harp data frame or harp list}

\item{member}{Does not need to be supplied if there is only one ensemble
member in \code{x}. Otherwise can be the member number, or the full name of
the member column to treat as deterministic.}

\item{sub_model}{For multimodel ensembles there may be columns with the
same member number. Set to the name of the sub-model to select \code{member} for
that sub-model.}
}
\value{
A harp deterministic data frame.
}
\description{
Convert an ensemble forecast to deterministic.
}
\examples{
as_det(ens_point_df, 0)
as_det(ens_grid_df, 1)
as_det(ens_point_list, 0)

# Name the member column explicitly
as_det(ens_point_df, "point_mbr000")
}
