% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{as_dttm}
\alias{as_dttm}
\title{Convert to date-time (POSIXct) object}
\usage{
as_dttm(x)
}
\arguments{
\item{x}{A numeric character string or numeric object}
}
\value{
A date-time object of class POSIXct
}
\description{
Given a string of numbers, a conversion is done to the POSIXct date-time
class. The string is assumed to be of the form YYYYMMDD, YYYYMMDDhh,
YYYYMMDDhhmm, or YYYYMMDDhhmmss with time zone UTC.
}
\examples{
# Character input
as_dttm("20220203")
as_dttm("2022020306")
as_dttm("202202030630")
as_dttm("20220203063022")
#
# Numeric input
as_dttm(20220203)
as_dttm(2022020306)
as_dttm(202202030630)
as_dttm(20220203063022)

}
