% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolist.R
\name{as_geolist}
\alias{as_geolist}
\alias{is_geolist}
\title{Convert a list of geofields to a geolist}
\usage{
as_geolist(...)

is_geolist(x)
}
\arguments{
\item{...}{list of geofields, or comma separated geofields.}

\item{x}{object to be tested}
}
\value{
A geolist
}
\description{
\code{geolists} are an integral data structure for harp. They allow for many
2d fields to be stored in columns in data frames so that other dimensions are
defined by other columns in the data frame. This avoids the need for
complicated dimensions for multidimensional data. A number of methods are
implemented for \code{geolists} enabling many elementwise mathematical
operations to be done efficiently.
}
\details{
All geofields in a geolist must be on the same domain, but they can have
different parameters, vertical levels, times etc.
}
\examples{
# Examples to come with the addition of test data
}
