% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\name{as_harp_df}
\alias{as_harp_df}
\title{Coerce to a harp_df data frame}
\usage{
as_harp_df(x)
}
\arguments{
\item{x}{a data frame with a \code{validdate} column.}
}
\value{
a data frame with the appropriate class
}
\description{
harp uses classes to dispatch data to the correct methods when running a
function. \code{as_harp_df} takes a data frame and scans the column names
and types to assign the correct classes. If the input data frame is not a
\link[tibble]{tibble} it will be coerced into one to ensure an easier to
read print method.
}
\examples{
library(dplyr)
d_f <- data.frame(validdate = as_datetime(seq_dates(2021010100, 2021010123)))
as_harp_df(d_f)
as_harp_df(mutate(d_f, fcst_det = runif(24)))
as_harp_df(mutate(d_f, fcst_mbr000 = runif(24), fcst_mbr001 = runif(24)))
# Note the class
class(as_harp_df(d_f))
class(as_harp_df(mutate(d_f, fcst_det = runif(24))))
class(as_harp_df(mutate(d_f, fcst_mbr000 = runif(24), fcst_mbr001 = runif(24))))
}
