% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\name{as_harp_list}
\alias{as_harp_list}
\title{Create a list of harp data frames}
\usage{
as_harp_list(...)
}
\arguments{
\item{...}{\code{harp_df} data frames. Must be named}
}
\value{
a \code{harp_list} list
}
\description{
Create a list of harp data frames
}
\examples{
as_harp_list(
  a = as_harp_df(data.frame(
    validdate = seq_dates(2021010100, 2021010123),
    a_det = runif(24)
  )),
  b = as_harp_df(data.frame(
    validdate = seq_dates(2021010100, 2021010123),
    b_det = runif(24)
  ))
)
as_harp_list(
  a = as_harp_df(data.frame(
    validdate = seq_dates(2021010100, 2021010123),
    a_mbr000  = runif(24),
    a_mbr001  = runif(24)
  )),
  b = as_harp_df(data.frame(
    validdate = seq_dates(2021010100, 2021010123),
    b_mbr000 = runif(24),
    b_mbr001 = runif(24)
  ))
)
}
