% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{as_str_dttm}
\alias{as_str_dttm}
\alias{as_YMD}
\alias{as_ymd}
\alias{as_YMDh}
\alias{as_ymdh}
\alias{as_YMDhm}
\alias{as_ymdhm}
\alias{as_YMDhms}
\alias{as_ymdhms}
\title{Convert to date-time string}
\usage{
as_str_dttm(x)

as_YMD(x)

as_ymd(x)

as_YMDh(x)

as_ymdh(x)

as_YMDhm(x)

as_ymdhm(x)

as_YMDhms(x)

as_ymdhms(x)
}
\arguments{
\item{x}{A POSIXct or numeric object in seconds since 1970-01-01 00:00:00}
}
\value{
A date-time string
}
\description{
These functions take a POSIXct or a numeric object and convert to a date-time
string with the format specified by the function. Years are always 4 digits
long with all other parts of the string 2 digits long with leading zeros
where necessary. In the case of numeric input, the number is assumed to be in
seconds since 1970-01-01 00:00:00 in UTC.
}
\details{
\code{as_str_dttm} returns a string with trailing zeros truncated. The
other functions are named such that the output is truncated to the precision
given by the function name. Note that no rounding is done, only truncation.
There are all lower case versions of functions as well as those with Y, M, D
in upper case.
}
\examples{
as_str_dttm(as.POSIXct("2022-03-08 00:00:00"))
as_str_dttm(as.POSIXct("2022-03-08 10:00:00"))
as_str_dttm(as.POSIXct("2022-03-08 10:30:00"))
as_str_dttm(as.POSIXct("2022-03-08 10:30:20"))
as_YMD(Sys.time())
as_YMDh(Sys.time())
as_YMDhm(Sys.time())
as_YMDhms(Sys.time())
as_YMD(as.numeric(Sys.time()))
as_YMDh(as.numeric(Sys.time()))
as_YMDhm(as.numeric(Sys.time()))
as_YMDhms(as.numeric(Sys.time()))
}
