% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{as_unixtime}
\alias{as_unixtime}
\title{Convert to UNIX time - seconds since 1970-01-01 00:00:00}
\usage{
as_unixtime(x)
}
\arguments{
\item{x}{A numeric character string, numeric or POSIXct object}
}
\value{
Numeric - seconds since 1970-01-01 00:00:00
}
\description{
Given a string of numbers or a POSIXct date-time object, a conversion is done
to UNIX time, i.e. seconds since 1970-01-01 00:00:00. The string is assumed
to be of the form YYYYMMDD, YYYYMMDDhh, YYYYMMDDhhmm, or YYYYMMDDhhmmss with
time zone UTC.
}
\examples{
# Character input
as_unixtime("20220203")
as_unixtime("2022020306")
as_unixtime("202202030630")
as_unixtime("20220203063022")
#
# Numeric input
as_unixtime(20220203)
as_unixtime(2022020306)
as_unixtime(202202030630)
as_unixtime(20220203063022)
#
# POSIXct input
as_unixtime(Sys.time())
}
