% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_stats.R
\name{circle_stats}
\alias{circle_stats}
\alias{circle_bias}
\alias{circle_mean}
\alias{circle_var}
\alias{circle_sd}
\alias{circle_sd_sq}
\title{Circular statistics}
\usage{
circle_bias(x, y, degrees = TRUE)

circle_mean(x, degrees = TRUE)

circle_var(x, degrees = TRUE)

circle_sd(x, degrees = TRUE)

circle_sd_sq(x, degrees = TRUE)
}
\arguments{
\item{x, y}{A numeric vector, matrix, or data frame in degrees or radians.}

\item{degrees}{Logical. Set to \code{TRUE} for x (and y) in degrees (the default).
If set to FALSE x (and y) will be assumed to be in radians.}
}
\value{
A numeric value or vector in the same units as the input.
}
\description{
These functions compute statistics for circular data. They would typically
be used when calculating statistics from wind direction. If \code{x} is a matrix
or a data frame, the statistics are computed for each row in the data.
}
\details{
Note that \code{circle_var()} gives a dimensionless output between 0 and 1,
whereas \code{circle_sd()} and \code{circle_sd_sq()} can be interpreted as being in
degrees (or radians depending on what units the data are in), or the square.
}
\examples{
# Doesn't cross zero
circle_bias(180, 90)
#
# Crosses zero
circle_bias(355, 5)
#
# Crosses zero the other way
circle_bias(5, 355)
# All in the 0 - 180 degrees range
circle_mean(c(20, 40, 60, 120, 140))
#
# Crossing zero
circle_mean(c(340, 30, 40, 320, 275))
# Clustered angles
circle_var(c(340, 10, 350, 5, 2, 355, 352, 8))
#
# Spread angles
circle_var(c(0, 90, 180, 270))
# Clustered angles
circle_sd(c(340, 10, 350, 5, 2, 355, 352, 8))
#
# Spread angles
circle_sd(c(0, 90, 180, 270))
# Clustered angles
circle_sd_sq(c(340, 10, 350, 5, 2, 355, 352, 8))
#
# Spread angles
circle_sd_sq(c(0, 90, 180, 270))
}
