% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{common_cases}
\alias{common_cases}
\title{Filter to common cases}
\usage{
common_cases(.fcst, ...)
}
\arguments{
\item{.fcst}{A harp_list object}

\item{...}{Extra columns from which to determine the common cases. To remove
one of the default columns from the test use -\if{html}{\out{<col>}}.}
}
\value{
The input data frame with only the common stations and forecast dates
for each forecast model selected.
}
\description{
For a fair comparison of models, the verification should only be done for
dates and locations that are common to all models. \code{common_cases} takes
a harp_list object as input and then identifies and filters to only those
cases that are common to all of the forecast models in the harp_list object.
By default this is done with the SID, fcst_dttm and lead_time columns, but
extra columns can be added via \code{...}. If one of the columns is a vertical
coordinate ("p", "z", "ml" for pressure, height and model level
respectively), that column will also be included.
}
