% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{deharp}
\alias{deharp}
\title{Remove harp classes}
\usage{
deharp(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{x} with all harp classes removed.
}
\description{
In some cases you may need to remove harp classes from a data frame, for
example if methods do not exist. Use this function to remove all harp related
classes from the object
}
\examples{
class(det_point_df)
class(deharp(det_point_df))
}
