% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_geolist.R
\name{domain_corners}
\alias{domain_corners}
\alias{domain_corners.geodomain}
\alias{domain_corners.geofield}
\alias{domain_corners.harp_geolist}
\title{Get the corners of a geodomain in longitude and latitude coordinates}
\usage{
domain_corners(x)

\method{domain_corners}{geodomain}(x)

\method{domain_corners}{geofield}(x)

\method{domain_corners}{harp_geolist}(x)
}
\arguments{
\item{x}{A geodomain, geofield or geolist.}
}
\value{
A named list containing the south-west (SW), north-east (NE),
south-east (SE) and north-west (NW) corners of the domain in longitude
and latitude coordinates.
}
\description{
Get the corners of a geodomain in longitude and latitude coordinates
}
\examples{
domain_corners(det_grid_df$fcst)

my_dom <- define_domain(10.9, 57.8, 100, 1, "longlat")
domain_corners(my_dom)
}
