% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_geolist.R
\name{domain_extent}
\alias{domain_extent}
\alias{domain_extent.geodomain}
\alias{domain_extent.geofield}
\alias{domain_extent.harp_geolist}
\title{Get the extent of a geodomain}
\usage{
domain_extent(x)

\method{domain_extent}{geodomain}(x)

\method{domain_extent}{geofield}(x)

\method{domain_extent}{harp_geolist}(x)
}
\arguments{
\item{x}{A geodomain, geofield or geolist.}
}
\value{
A named list containing:
\item{lonlim,latlim}{The extreme longitudes and latitudes in the domain}
\item{clonlat}{The longitude and latitude of the domain centre}
\item{x0,x1,y0,y1}{The SW and NE corners of the domain in projection coordinates}
\item{dx,dy}{The spacing between grid points in projection coordinates}
\item{nx,ny}{The number of grid points in the x and y directions}
}
\description{
Get the extent of a geodomain
}
\examples{
domain_extent(det_grid_df$fcst)

my_dom <- define_domain(10.9, 57.8, 100, 1, "longlat")
domain_extent(my_dom)
}
