% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods.R
\name{ens_prob}
\alias{ens_prob}
\title{Title}
\usage{
ens_prob(
  x,
  threshold = 0,
  comparator = c("ge", "gt", "le", "lt", "between", "outside"),
  include_low = TRUE,
  include_high = TRUE,
  ...
)
}
\arguments{
\item{include_high}{}
}
\description{
Title
}
