% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{ens_stats}
\alias{ens_stats}
\title{Title}
\usage{
ens_stats(
  .fcst,
  mean = TRUE,
  sd = TRUE,
  var = FALSE,
  min = FALSE,
  max = FALSE,
  keep_members = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
