% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{expand_date}
\alias{expand_date}
\title{Expand date-time column in a data frame}
\usage{
expand_date(.data, col, text_months = FALSE)
}
\arguments{
\item{.data}{A data frame or harp_list of data frames}

\item{col}{The name of the date-time to column to be expanded. Can be quoted
or unquoted. If using a variable, it should be wrapped in {{.}}}

\item{text_months}{Logical. If TRUE, month names are used rather than numbers.}
}
\value{
A data frame or harp_list of data frames with new columns for the
expanded date
}
\description{
\code{expand_date} extracts the year, month, day, hour and minute from
a date-time column in a data frame and creates a column for each part
of the date. If the name of the date-time column ends with "date", the names
of the new columns are whatever precedes "date" in the original column name
followed by "_year", "_month", "_day", "_hour", "_minute". If the original
column name does not end with date the suffixes are pasted onto the
original column name.
}
\examples{
expand_date(det_point_df, fcdate)
expand_date(det_point_list, validdate)
expand_date(ens_point_df, validdate, text_months = TRUE)

# Column name can be quoted
expand_date(ens_grid_df, "fcdate")

# If using a variable, wrap in {{<var>}}
my_col <- "fcdate"
expand_date(ens_grid_df, {{my_col}})
}
