% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract_numeric}
\alias{extract_numeric}
\title{Extract numeric values from a string or vector of strings}
\usage{
extract_numeric(x, as_list = FALSE)
}
\arguments{
\item{x}{A character vector.}

\item{as_list}{Logical. Whether to return a list. See Details.}
}
\value{
A vector or list of numerics
}
\description{
If you have a string that contains characters and numbers,
\code{extract_numeric} will extract the numeric parts. Anything between the
numerics in the string will result in a separate element in the output.
}
\details{
If a vector of strings is passed and you want to return something of the
same length, set \code{delist = FALSE} and a list will be returned with the
same length as the input vector.
}
\examples{
extract_numeric("3h")

# With decimals
extract_numeric("3h 4h 5.5h")

# With negative values
extract_numeric("-2s4s 6s")

# With a vector of values
extract_numeric(c("2h", "3h", "7h"))
extract_numeric(c("2h", "3h 4h 5h 6h", "7h"))

# Return a list of the same length as the vector
extract_numeric(c("2h", "3h 4h 5h 6h", "7h"), as_list = TRUE)
}
