% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{filter.harp_list}
\alias{filter.harp_list}
\title{Subset rows using column values}
\usage{
\method{filter}{harp_list}(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Expressions that return a
logical value, and are defined in terms of the variables in \code{.data}.
If multiple expressions are included, they are combined with the \code{&} operator.
Only rows for which all conditions evaluate to \code{TRUE} are kept.}
}
\description{
This is the \code{harp_list} method for \code{\link[dplyr]{filter}()}. It
works in exactly the same way except that a \code{harp_list} is returned.
To use this function, the \code{dplyr} package must be attached and the
\code{.harp_list} suffix can be dropped.
}
