% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{from_seconds}
\alias{from_seconds}
\alias{fix_seconds_names}
\alias{add_seconds_names}
\title{Convert seconds to another time unit}
\usage{
from_seconds(x, to = "h", as_names = FALSE, from_names = FALSE, num_dp = NULL)

fix_seconds_names(x)

add_seconds_names(x, to = "h")
}
\arguments{
\item{x}{A vector of numeric values in seconds.}

\item{to}{The unit to convert from seconds to. Can be "s", "m", "h", "d", "w"
for seconds, minutes, hours, days and weeks respectively. The default is to
convert to hours.}

\item{as_names}{Logical. Set to \code{TRUE} to return \code{x} unchanged, but with the
values in the new units as names. Default is \code{FALSE}}

\item{from_names}{Logical. Set to \code{TRUE} to get the time unit to convert to
from the names of \code{x}. In this case \code{to} is ignored. Default is FALSE.}

\item{num_dp}{If non integer values are returned, the number of decimal
places to include in the values in the new units. Set to \code{NULL} (the
default) to format "as is".}
}
\value{
A character vector or a named numeric vector.
}
\description{
Given a vector of values in seconds, \code{from_seconds()} will return the
vector in the specified time unit. Values will be a character vector with
each element appended with the unit, "s" = seconds, "m" = minutes, "h" =
hours, "d" = days and "w" = weeks. Alternatively the function can return the
values unchanged, but with names added in the specified unit. If the input
already has names, the required unit will be inferred from those names.
}
\details{
There are also a couple of convenience functions that act as wrappers
around \code{from_seconds()} - \code{add_seconds_names()} simply adds names in the
specified unit to the input, ignoring any names that are already there, and
\code{fix_seconds_names()} ensures that the names are correct with the unit
implied from the names that are already there. This is useful if some
mathematical operation has been done and the names need updating.
}
\examples{

from_seconds(c(0, 3600, 7200, 10800))
from_seconds(c(0, 3600, 7200, 10800), as_names = TRUE)

# Get time unit from names
tt <- to_seconds(seq_hours(0, 9, 3))
tt
from_seconds(tt, from_names = TRUE)

# Fix the names after mathematical operation
tt + 3600 # Wrong names!
fix_seconds_names(tt + 3600) # Correct names
}
\seealso{
\code{\link[=to_seconds]{to_seconds()}}
}
