% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_transform.R
\name{geo_points}
\alias{geo_points}
\title{Title}
\usage{
geo_points(
  x,
  points,
  method = c("bilinear", "nearest", "bicubic"),
  mask = NULL,
  force = FALSE,
  weights = NULL,
  keep_weights = FALSE
)
}
\arguments{
\item{keep_weights}{}
}
\description{
Title
}
