% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_transform.R
\name{geo_regrid}
\alias{geo_regrid}
\title{Title}
\usage{
geo_regrid(
  x,
  new_grid,
  method = c("bilinear", "nearest", "bicubic"),
  mask = NULL,
  new_mask = NULL,
  weights = NULL,
  keep_weights = FALSE
)
}
\arguments{
\item{keep_weights}{}
}
\description{
Title
}
