% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_transform.R
\name{geo_reproject}
\alias{geo_reproject}
\title{Reproject from or to lat-lon coordinates}
\usage{
geo_reproject(
  x,
  proj,
  x_col = "lon",
  y_col = "lat",
  crop = FALSE,
  inverse = FALSE
)
}
\arguments{
\item{x}{A data frame}

\item{proj}{The projection. Can be a \code{geodomain}, a \code{geofield}, a \code{geolist},
a projection string or a meteogrid projection list. When \code{inverse = FALSE},
this is the projection to which locations in lat-lon coordinates are
reprojected, and when \code{inverse = TRUE}, this the projection from which
locations in projected coordinates are reprojected to lat-lon coordinates.}

\item{x_col, y_col}{The names of the columns containing the x and y
coordinates to be reprojected. For \code{inverse = FALSE}, these should be
longitude and latitude in decimal degrees. For \code{inverse = TRUE}, these
should be eastings and northings in metres.}

\item{crop}{When \code{proj} is a \code{geodomain}, \code{geofield} or \code{geolist}, set to
\code{TRUE} to crop the reprojected locations to only those locations within the
domain.}

\item{inverse}{Set to \code{TRUE} to reprojected from projected coordinates to
lat-lon coordinates. The default is \code{FALSE}}
}
\value{
The input data frame with new columns for the reprojected
coordinates. The projection is added as an attribute. If the data are
cropped, the domain is also added.
}
\description{
Reproject from or to lat-lon coordinates
}
\examples{
geo_reproject(station_list, det_grid_df$fcst)

# Crop to domain
geo_reproject(station_list, det_grid_df$fcst, crop = TRUE)

# inverse projection
projected <- geo_reproject(station_list, det_grid_df$fcst)
geo_reproject(
  projected, det_grid_df$fcst, x_col = x, y_col = y, inverse = TRUE
)

}
