% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_transform.R
\name{geo_transform}
\alias{geo_transform}
\title{Title}
\usage{
geo_transform(
  x,
  trans = c("points", "regrid", "subgrid", "zoom", "xsection"),
  opts
)
}
\arguments{
\item{opts}{}
}
\description{
Title
}
