% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_weights.R
\name{geo_weights_regrid}
\alias{geo_weights_regrid}
\title{Title}
\usage{
geo_weights_regrid(
  x,
  new_grid,
  method = c("bilinear", "nearest", "bicubic"),
  mask = NULL,
  new_mask = NULL
)
}
\arguments{
\item{new_mask}{}
}
\description{
Title
}
