% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_weights.R
\name{geo_weights_xsection}
\alias{geo_weights_xsection}
\title{Title}
\usage{
geo_weights_xsection(
  x,
  p1,
  p2,
  n = 100,
  method = c("bilinear", "nearest", "bicubic")
)
}
\arguments{
\item{method}{}
}
\description{
Title
}
