% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_transform.R
\name{geo_xsection}
\alias{geo_xsection}
\title{Title}
\usage{
geo_xsection(
  x,
  p1,
  p2,
  n = 100,
  method = c("bilinear", "nearest", "bicubic"),
  weights = NULL,
  keep_weights = FALSE
)
}
\arguments{
\item{keep_weights}{}
}
\description{
Title
}
