% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_functions.R
\name{geopotential_to_height}
\alias{geopotential_to_height}
\title{Convert geopotential to height}
\usage{
geopotential_to_height(x)
}
\arguments{
\item{x}{}
}
\value{
The same class as x
}
\description{
A simple conversion of geopotential in m^2 / s^2 to height in m, by dividing
by acceleration due to gravity (g = 9.80665).
}
