% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_geolist.R
\name{get_domain}
\alias{get_domain}
\title{Extract domain information as a geodomain object}
\usage{
get_domain(x)
}
\arguments{
\item{x}{A geofield or geolist}
}
\value{
A geodomain
}
\description{
The "domain" attribute is extracted from the input and returned if it is a
geodomain.
}
\examples{
get_domain(det_grid_df$fcst)

get_domain(det_grid_df$fcst[[1]])
}
