% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_geolist.R
\name{glapply2}
\alias{glapply2}
\title{Apply a function to element pairs of 2 geolists}
\usage{
glapply2(X, Y, FUN, ...)
}
\arguments{
\item{X}{A geolist}

\item{Y}{A geolist of the same length as \code{X} that is on the same domain.}

\item{FUN}{The function to apply to \code{X} and \code{Y}.}

\item{...}{Other arguments to \code{FUN}.}
}
\value{
A \code{geolist} of the same length as \code{X} and \code{Y}.
}
\description{
\code{glapply2()} is variant of \code{glapply()} whereby a function that takes 2
\code{geofield}s as arguments is applied element-wise to 2 geolists. It is
equivalent to \code{\link[purrr]{map2}}, but is explicitly for \code{geolist}s.
}
\examples{
glapply2(
  ens_grid_df$grid_mbr000, ens_grid_df$grid_mbr001, function(x, y) x + y
)
}
