% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_geolist.R
\name{gmapply}
\alias{gmapply}
\title{Apply a funtion that returns a geolist to mutiple list are vector arguments}
\usage{
gmapply(FUN, ..., MoreArgs = NULL)
}
\arguments{
\item{FUN}{function to apply, found via \code{\link[base]{match.fun}}.}

\item{...}{arguments to vectorize over, will be recycled to common
  length (zero if one of them is).  See also \sQuote{Details}.}

\item{MoreArgs}{a list of other arguments to \code{FUN}.}
}
\value{
A \code{geolist} of the same length as the arguments in \code{...}
}
\description{
\code{gmapply()} is an implementation of \code{\link[base]{mapply}()} especially
for geolists. \code{FUN} is applied to the first elements of each \code{...} argument,
the second elements, the third elements and so on. \code{gmapply()} differs from
\code{mapply()} in that the result is expected to be a \code{geolist}.
}
\examples{
gmapply(
  function(a, b, c, d) (a + b * c) / d,
  det_grid_df$fcst,
  seq_len(nrow(det_grid_df)),
  seq_len(nrow(det_grid_df)) * 10,
  MoreArgs = list(d = 100)
)
}
