% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{hybrid_xs_to_grid}
\alias{hybrid_xs_to_grid}
\title{Title}
\usage{
hybrid_xs_to_grid(
  x,
  psfc,
  level_defs = harmonie_levels$full65,
  temp = NA,
  spec_hum = NA,
  topo = NA,
  topo_is_geo = TRUE,
  vert_res = "auto",
  out_of_bounds = c("censor", "squish")
)
}
\arguments{
\item{out_of_bounds}{}
}
\description{
Title
}
