% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{join_to_fcst}
\alias{join_to_fcst}
\title{Join data to a forecast}
\usage{
join_to_fcst(
  .fcst,
  .join,
  join_type = c("inner", "left", "right", "full", "semi", "anti"),
  by = NULL,
  latlon = FALSE,
  elev = FALSE,
  force = FALSE,
  keep_x = TRUE,
  keep_y = FALSE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_df} data frame or a \code{harp_list}.}

\item{.join}{A data frame to join to the forecast.}

\item{join_type}{How to join the data frame. Acceptable values are: "inner",
"left", "right", "full", "semi", "anti". See \code{\link[dplyr]{join}} for
more details.}

\item{by}{Which columns to join by - if set to NULL a natural join will be
done, using all variables with common names across .fcst and .join. The
default is to join using all common columns in .fcst and .join excluding
lat, lon and elev. This is because they may be stored to different levels of
precision and the join will thus fail.}

\item{latlon}{Logical. Whether to include latitude and longitude columns in
the default for \code{by}. The default is FALSE.}

\item{elev}{Logical. Whether to include the station elevation column in
the default for \code{by}. The default is FALSE.}

\item{force}{Set to TRUE to force the join to happen even if the units
in .fcst and .join are not compatible.}

\item{keep_x, keep_y}{Where duplicate column names are found, but not used in
the join, these arguments are used to indicate whether the duplicate
columns from .fcst (\code{keep_x}), or .join (\code{keep_y}) should be
kept. The default is \code{keep_x = TRUE, keep_y = FALSE}.}

\item{...}{Other arguments for \link[dplyr]{join}.}
}
\value{
The input forecast data frame with column(s) added from \code{.join}.
}
\description{
\code{join_to_fcst} is a special case of the join family of functions. It's
primary purpose is to join data frame of observations to a data frame or
harp_list of forecasts such that forecast - observation pairs are on the
same row in the joined data frame. An extra check is made to make sure that
the forecast data and observations data are in the same units.
}
\examples{
# Make some fake observations
library(tibble)
obs <- tibble(
  valid_dttm = det_point_df$valid_dttm,
  SID       = det_point_df$SID,
  units     = "degC",
  T2m       = runif(nrow(det_point_df))
)

# Make sure the forecast has units
fcst <- set_units(det_point_df, "degC")

join_to_fcst(fcst, obs)

# Also works for harp_list objects
join_to_fcst(set_units(det_point_list, "degC"), obs)

# And works with gridded data
join_to_fcst(set_units(ens_grid_df, "degC"), set_units(anl_grid_df, "degC"))
}
