% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_transform.R
\name{make_domain}
\alias{make_domain}
\title{Title}
\usage{
make_domain(
  centre_lon,
  centre_lat,
  nxny,
  dxdy,
  proj = c("lambert", "lcc", "merc", "mercator", "omerc", "tmerc", "somerc", "lalo",
    "longlat", "latlong", "ob_tran", "rot_longlat", "rot_latlong", "RotLatLon", "stere",
    "stereo", "stereographic"),
  ref_lon,
  ref_lat,
  exey = NULL,
  tilt = 0,
  R = 6371229,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
