% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{member_colnames}
\alias{member_colnames}
\title{Retrieve column names for ensemble members}
\usage{
member_colnames(x, exclude = NULL, invert = FALSE)
}
\arguments{
\item{x}{A \code{harp_ens_point_df}, or a \code{harp_ens_grid_df} data frame, or a
character vector of column names.}

\item{exclude}{The members to exclude from the output. Note that if numeric,
lagged member names with the same member number will also be excluded. To
avoid this, a string or regular expression could be passed.}

\item{invert}{Logical. Whether to invert the output. That is, if exclude has
a value, output only those column names rather than exclude them.}
}
\value{
A character vector of column names.
}
\description{
Given a \code{harp_ens_point_df} or a \code{harp_ens_grid_df} data frame, or a
character vector, the names of the columns containing data for ensemble
members will be retrieved. This is done by searching for the regular
expression "_mbr[0-9]{3}", which is the standard used in harp for naming
ensemble members. Members can be excluded using the \code{exclude} argument.
}
\examples{
member_colnames(ens_point_df)

# Exclude member 0
member_colnames(ens_point_df, exclude = 0)

# Only member 0
member_colnames(ens_point_df, exclude = 0, invert = TRUE)

# All members with the same number will be exluded for lagged ensembles
lagged_ens <- ens_point_df
lagged_ens$point_mbr000_lag1h = runif(nrow(ens_point_df))
colnames(lagged_ens)
member_colnames(lagged_ens, exclude = 0)

# Use a regular expression to not exclude the lagged members
member_colnames(lagged_ens, exclude = "_mbr000$")
}
