% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_functions.R
\name{met_functions}
\alias{met_functions}
\alias{rh_to_td}
\alias{q_to_rh}
\title{Compute dewpoint temperature from temperature and relative humidity}
\usage{
rh_to_td(rh, t, a = 17.67, b = 243.5)

q_to_rh(q, t, p = 1013.25, a = 17.67, b = 243.5)
}
\arguments{
\item{rh}{A vector of relative humidities expressed as decimal fractions
between 0 and 1. If the maximum is greater than 5, it is assumed that all
relative humidities are in \eqn{\%} and will be divided by 100.}

\item{t}{A vector of the same length as \code{rh} temperatures in
\eqn{\degree C}. If the maximum is greater than 200, it is assumed that
all temperatures are in \eqn{Kelvin} and they will be converted to to
\eqn{\degree C}.}

\item{a}{A tuning parameter. The default value is \eqn{17.67}.}

\item{b}{A tuning parameter. The default value is \eqn{243.5}.}

\item{q}{The specific humidity in \eqn{kg.kg^{-1}}}

\item{p}{The air pressure in \eqn{hPa}}
}
\value{
A numeric vector of the same length as \code{t}.
}
\description{
The computation of dewpoint temperature is done using the formula
\deqn{\frac{b\left(ln(RH) + \left(\frac{aT}{(b+T)}\right)\right)}{\left(a-\left(ln(RH) + \left(\frac{aT}{(b+T)}\right)\right)\right)}}
And the computation for relative humidity is done following
\deqn{RH=\frac{e}{e_s}}
where
\deqn{e=\frac{QP}{0.378Q+0.622}}
\deqn{e_s=6.112e^{\frac{aT}{b+T}}}
In the above formulae:\cr
\eqn{T} is the temperature in \eqn{\degree C} \cr
\eqn{RH} is the relative humidity expressed as a decimal between 0 and 1 \cr
\eqn{Q} is the specific humidity in \eqn{kg.kg^{-1}} \cr
\eqn{P} is the atmospheric pressure in \eqn{hPa} \cr
\eqn{a} and \eqn{b} are tuning parameters \cr
\cr
The default values for \eqn{a} and \eqn{b} are 17.67 and 243.5 respectively as
recommended by NOAA.
}
\details{
Other tunings for \eqn{a} and \eqn{b} are possible:

\eqn{a=17.27}, \eqn{b=237.3} is valid for \eqn{0\lt T \lt 60}

Alternatively:

\eqn{a=17.62}, \eqn{b=243.12}\cr
\eqn{a=17.625}, \eqn{b=243.04}

Finally A.L. Buck (1981) recommends different tunings based on
temperature:

\eqn{a=17.368}, \eqn{b=238.88} for \eqn{T \gt 0}\cr
\eqn{a=17.966}, \eqn{b=247.15} for \eqn{T \lt 0}\cr

Note that the formulation used for dewpoint temperature does not take
atmospheric pressure into account, so may not be correct at higher
altitudes.
}
\examples{
rh_to_td(0.8, 20)

# In Kelvin and \%
rh_to_td(80, 293.15)
#
# Relative humidity is returned as a decimal fraction between 0 and 1
q_to_rh(0.013, 20)
}
