% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parse_thresholds}
\alias{parse_thresholds}
\title{Decide whether thresholds are absolute or quantiles}
\usage{
parse_thresholds(x)
}
\arguments{
\item{x}{A numeric or character vector.}
}
\value{
A list with a numeric vector of thresholds and a logical indicating
whether the thresholds are quantiles.
}
\description{
For quantiles, thresholds should be expressed as a character vector with
each element beginning with "q" followed by the quantile expressed as a
value between 0 and 1.
}
\examples{
parse_thresholds(seq(1, 5))
parse_thresholds(paste0("q", seq(0, 1, 0.2)))
}
