% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{pivot_parameters}
\alias{pivot_parameters}
\alias{pivot_parameters_wider}
\alias{pivot_parameters_longer}
\title{Title}
\usage{
pivot_parameters_wider(x, values_from)

pivot_parameters_longer(x, cols, values_to)
}
\arguments{
\item{values_from}{}
}
\description{
Title
}
