% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_functions.R
\name{potential_temp}
\alias{potential_temp}
\title{Potential temperature}
\usage{
potential_temp(temp, press, press_ref = 1e+05)
}
\arguments{
\item{temp}{Temperature (K)}

\item{press}{Air parcel pressure (Pa)}

\item{press_ref}{Reference pressure (Pa). The default is 100000 Pa.}
}
\value{
Potential temperature (K)
}
\description{
Compute the potential temperature of an air parcel - that is the temperature
it would have if it were brought adiabatically to a reference temperature.

The potential temperature is defined as:
\deqn{ \theta = T \left( \frac{p_0}{p} \right)^{R/c_p} }

where \eqn{\theta} is potential temperature, \eqn{T} is temperature,
\eqn{p} is pressure, \eqn{p_0} is reference pressure, \eqn{R} is the
gas constant for dry air, and \eqn{c_p} is the specific heat at
constant pressure.
}
\examples{
potential_temp(280, 85000)
}
