% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{pull.harp_list}
\alias{pull.harp_list}
\title{Extract a single column}
\usage{
\method{pull}{harp_list}(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{For use by methods.}
}
\description{
This is the \code{harp_list} method for \code{\link[dplyr]{pull}()}. It
works in exactly the same way except that a list of vectors is returned.
To use this function, the \code{dplyr} package must be attached and the
\code{.harp_list} suffix can be dropped.
}
