% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{select.harp_list}
\alias{select.harp_list}
\title{Subset columns using their names and types}
\usage{
\method{select}{harp_list}(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}
}
\description{
This is the \code{harp_list} method for \code{\link[dplyr]{select}()}. It
works in exactly the same way except that a \code{harp_list} is returned.
To use this function, the \code{dplyr} package must be attached and the
\code{.harp_list} suffix can be dropped.
}
