% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{select_members}
\alias{select_members}
\title{Select ensemble members}
\usage{
select_members(.data, members, include_lagged = TRUE)
}
\arguments{
\item{.data}{A data frame with ensemble members or a \code{harp_list}}

\item{members}{The members to select. Can be a numeric vector, or a named
list to select members from specific forecast models in a \code{harp_list}
object.}

\item{include_lagged}{Logical. Whether to include lagged ensemble members
in the selection.}
}
\value{
A data frame with the selected members or a \code{harp_list} of
data frames with the selected ensemble members.
}
\description{
\code{select_members} is used to select specific ensemble members from a
data frame along with all other columns. The method can also be applied to
a \code{harp_list} of ensemble data frames.
}
\examples{
select_members(ens_point_df, 0)
select_members(ens_grid_df, 1)

# More than one member can be selected
select_members(ens_point_df, c(0, 1))

# Select member 0 from a harp_list
select_members(ens_point_list, 0)

# Different members can be selected from each data frame
select_members(ens_point_list, list(a = 0, b = 1))
}
