% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{seq_double}
\alias{seq_double}
\title{Generate a doubled sequence}
\usage{
seq_double(start, len)
}
\arguments{
\item{start}{The first number in the sequence.}

\item{len}{The length of the sequence.}
}
\value{
A numeric vector
}
\description{
\code{seq_double} generates a sequence of numbers where each value is double
the previous value in the sequence.
}
\examples{
seq_double(1, 5)
seq_double(0.125, 10)
}
