% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{seq_pwr2}
\alias{seq_pwr2}
\title{Generate a sequence of integer powers of 2}
\usage{
seq_pwr2(from, len, powers = TRUE)
}
\arguments{
\item{from}{The starting power or the value of the first number in the
sequence. In the case of the latter, this must be value that is 2
raised to an integer power.}

\item{len}{The length of the desired sequence.}

\item{powers}{Logical. Whether the first argument is the first power to
begin the sequence from, or whether it is the value to begin the sequence
from.}
}
\value{
A numeric vector of consecutive powers of 2.
}
\description{
A sequence of powers of 2 is generated based on the first value and the
length of the desired sequence. The first value could either be the first
power in the sequence, or it could be the first value of the sequence. The
\code{powers} argument is used to tell the function what the first value is.
}
\details{
The function only generates sequences of 2 raised to integer powers. It can
be useful in generating breaks values for exponential colour scales.
}
\examples{
# Start from 1
seq_pwr2(0, 5)
seq_pwr2(1, 5, powers = FALSE)
# Start from a value larger than 1
seq_pwr2(3, 5)
seq_pwr2(8, 5, powers = FALSE)
# Start from a value smaller than 1
seq_pwr2(-3, 5)
seq_pwr2(0.125, 5, powers = FALSE)
}
