% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{set_units}
\alias{set_units}
\title{Add or modify a units column}
\usage{
set_units(x, units)
}
\arguments{
\item{x}{A data frame or a harp_list.}

\item{units}{The units name to put in the units column.}
}
\value{
A data frame or harp_list of the same size as x with the units
column either added or modified.
}
\description{
Add or modify a units column
}
\examples{
# det_point_df has no units column
det_point_df

# Set units to "degC"
new_det_point_df <- set_units(det_point_df, "degC")
new_det_point_df

# Modify the units name to "degrees_C"
set_units(new_det_point_df, "degrees_C")
}
