% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{to_seconds}
\alias{to_seconds}
\title{Convert a time period to seconds}
\usage{
to_seconds(x)
}
\arguments{
\item{x}{A character or numeric vector. The units of the input is specified
by a single character:
\itemize{
\item{s} {seconds}
\item{m} {minutes}
\item{h} {hours}
\item{d} {days}
\item{w} {weeks}
}}
}
\value{
A named vector where the values are the number of seconds and the
names are the equivalent in the units given in the input
}
\description{
Convert number of minutes, hours, days or weeks into seconds. If \code{x} is
numeric, it is assumed to be in hours.
}
\examples{
# Numeric inputs are assumed to be in hours
to_seconds(c(0, 1, 2))

# The same values given in seconds, minutes and explicitly hours
to_seconds(c("0s", "3600s", "7200s"))
to_seconds(c("0m", "60m", "120m"))
to_seconds(c("0h", "1h", "2h"))

# Units can be mixed
to_seconds(c(paste0(24 * 7, "h"), "7d", "1w"))
}
