% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{unixtime_to_datetime}
\alias{unixtime_to_datetime}
\alias{unixtime_to_YMD}
\alias{unixtime_to_YMDh}
\alias{unixtime_to_YMDhm}
\alias{unixtime_to_YMDhms}
\alias{unixtime_to_ymd}
\alias{unixtime_to_ymdh}
\alias{unixtime_to_ymdhm}
\alias{unixtime_to_ymdhms}
\alias{unixtime_to_str_datetime}
\title{Convert unix time to other formats.}
\usage{
unixtime_to_datetime(x)

unixtime_to_YMD(x)

unixtime_to_YMDh(x)

unixtime_to_YMDhm(x)

unixtime_to_YMDhms(x)

unixtime_to_ymd(x)

unixtime_to_ymdh(x)

unixtime_to_ymdhm(x)

unixtime_to_ymdhms(x)

unixtime_to_str_datetime(x)
}
\arguments{
\item{x}{A numeric vector of UNIX time in seconds since 1970-01-01 00:00:00.}
}
\value{
A date-time object or a character date-time string in the desired
format
}
\description{
The unix time is expected to be in seconds since 1970-01-01 00:00:00 for
the UTC time zone.
}
\examples{
unixtime_to_datetime(Sys.time())
unixtime_to_ymd(Sys.time())
unixtime_to_ymdh(Sys.time())
unixtime_to_ymdhm(Sys.time())
unixtime_to_ymdhms(Sys.time())
unixtime_to_str_datetime(Sys.time())
}
