% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolist.R
\name{variance}
\alias{variance}
\alias{std_dev}
\title{Variance and standard deviation}
\usage{
variance(x, na.rm = FALSE, ...)

std_dev(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{a geolist, numeric vector, matrix or data frame.}

\item{na.rm}{logical. Should missing values be removed?}

\item{...}{Other arguments to \link[stats]{sd} or \link[stats]{var} for non geolists.}
}
\value{
If x is a geolist, a geofield.
}
\description{
\code{\link[stats]{sd}} and \code{\link[stats]{var}} are not implemented
as methods, which means thet cannot be dispatched for geolists. \code{variance}
and \code{std_dev} are implemented as methods and can be used to compute
the standard deviation and variance of geolists in a pointwise manner.
If the input is not a geolist the default functions \code{\link[stats]{sd}} and
\code{\link[stats]{var}} are used.
}
