% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_functions.R
\name{virtual_temp}
\alias{virtual_temp}
\title{Compute virtual temperature}
\usage{
virtual_temp(temp, spec_hum)
}
\arguments{
\item{temp}{Air temperature in K.}

\item{spec_hum}{Specific humidity in kg/kg.}
}
\value{
Virtual temperature in K.
}
\description{
The virtual temperature is computed from the air temperature and the
specific humidity of the air using the formula \verb{Tv = T(1 + 0.61Q)}. If the
function suspects that the inputs are in the wrong units (degrees C instead
of Kelvin; g/kg instead of kg/kg), they are converted to the correct units
with a warning.
}
\examples{
virtual_temp(293.15, 0.01)

# Corrections to inputs made if in wrong units
virtual_temp(20, 10)
}
