% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_functions.R
\name{vorticity}
\alias{vorticity}
\alias{divergence}
\title{Compute the vorticity or divergence from vector components}
\usage{
vorticity(u, v, rotate_to_grid = FALSE, rotation_init = NULL)

divergence(u, v, rotate_to_grid = FALSE, rotation_init = NULL)
}
\arguments{
\item{u}{A geofield or geolist of the u component}

\item{v}{A geofield or geolist of the v component}

\item{rotate_to_grid}{Logical. Whether to rotate the vector components from
earth relative to grid relative. Default is FALSE}

\item{rotation_init}{If the angles for rotation have been previously
computed (e.g. using \code{meteogrid::geowind.init}) they can be passed here.
Default is NULL}
}
\value{
A geofield if \code{u} and \code{v} are geofields or a geolist if \code{u} and \code{v}
are geolists.
}
\description{
These functions compute the vorticity and divergence of a vector field. If
the vectors are relative to the earth but the grid is on a projection, the
vectors can optionally be rotated to the projected grid prior to computation.
}
