% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_transform.R
\name{define_domain}
\alias{define_domain}
\title{Define a geodomain}
\usage{
define_domain(
  centre_lon,
  centre_lat,
  nxny,
  dxdy,
  proj = c("lambert", "lcc", "merc", "mercator", "omerc", "tmerc", "somerc", "lalo",
    "longlat", "latlong", "ob_tran", "rot_longlat", "rot_latlong", "RotLatLon", "stere",
    "stereo", "stereographic"),
  ref_lon,
  ref_lat,
  exey = NULL,
  tilt = 0,
  R = 6371229,
  ...
)
}
\arguments{
\item{centre_lon}{The longitude of the centre of the domain in decimal
degrees.}

\item{centre_lat}{The latitude of the centre of the domain in decimal
degrees.}

\item{nxny}{The number of grid points in the x and y directions. Should be a
vector of length 2 with the number of grid points in the x direction first.
If only 1 value is given it is assumed to be the same in both directions.}

\item{dxdy}{The horizontal resolution of the grid in the x and y directions.
For lat-lon projections this should be in decimal degrees, otherwise should
be in metres. Should be a vector of length 2 with the resolution in the x
direction first.}

\item{proj}{The projection of the domain. See details.}

\item{ref_lon}{The reference longitude of the projection if needed.}

\item{ref_lat}{The reference latitude of the projection if needed.}

\item{exey}{If defining a grid with an extension zone, a vector length 2 with
the number of grid points in the x and y directions of the extension zone.
If only 1 value is given it is assumed to be the same in both directions.}

\item{tilt}{The tilt used in a rotated Mercator projection.}

\item{R}{The radius of the earth in metres. The default is 6371229m.}

\item{...}{Other arguments describing the shape of the earth in \code{proj}
format.}
}
\value{
A \code{geodomain}
}
\description{
This function is used to define a new domain with a regular grid. At a
minimum, the projection, geographic location of the centre of the domain and
number of horizontal resolution of the grid points must be provided.
}
\examples{
dd <- define_domain(10, 60, 1000, 2500) # Default lambert projection
plot(dd)

dd <- define_domain(0, 0, c(360, 180), 1, "latlong") # Whole earth
plot(dd)
}
