% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{ens_prob}
\alias{ens_prob}
\alias{ens_prob.harp_ens_point_df}
\title{Compute the ensemble probability for a threshold}
\usage{
ens_prob(
  x,
  threshold = 0,
  comparator = c("ge", "gt", "le", "lt", "between", "outside"),
  include_low = TRUE,
  include_high = TRUE,
  ...
)

\method{ens_prob}{harp_ens_point_df}(
  x,
  threshold = 0,
  comparator = c("ge", "gt", "le", "lt", "between", "outside"),
  include_low = TRUE,
  include_high = TRUE,
  keep_members = FALSE,
  obs_col = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{harp_ens_grid_df} or \code{harp_ens_point_df} data frame, a \code{geolist},
or a \code{harp_list} containing ensemble data frames.}

\item{threshold}{A threshold for which to compute probabilities. If
\code{comparator = "between"} or \code{comparator = "outside"}, it must be a vector
of length 2. For \code{harp_ens_point_df} data frames, this can also be a
character type with numeric values between 0 and 1 preceded by "q" to
denote quantile thresholds.}

\item{comparator}{How to compare \code{x} with the threshold to compute
binary probabilities. Can be \code{"ge"}, \code{"gt"}, \code{"le"}, or
\code{"lt"} for >=, >, <= and < respectively. Can also be
\code{"between"} or \code{"outside"}, in which case the binary probability
of being between or outside of the two values given in \code{threshold} is
computed.}

\item{include_low}{Logical. Whether to include to the lower of the two
thresholds in the comparison when \code{comparator = "between"} or
\code{comparator = "outside"}.}

\item{include_high}{Logical. Whether to include to the higher of the two
thresholds in the comparison when \code{comparator = "between"} or
\code{comparator = "outside"}.}

\item{...}{Used for methods.}

\item{keep_members}{Logical. Whether to keep the ensemble member data after
computing the probabilities.}

\item{obs_col}{The name of the column containing observations. If not NULL,
binary probabilities are computed for this column. Can be the name of the
column quoted or unquoted, or if a variable it must be embraced in double
curly brackets \code{{{ }}}. Must be supplied if thresholds are quantiles.}
}
\value{
An object of the same class as \code{x} with the probabilities computed
}
\description{
The probability for a threshold for an ensemble is computed. This is by
default for threshold exceedence (P(fcst >= threshold)), but the probability
for being below the threshold, or between or outside of two thresholds can
also be calculated.
}
\details{
Note that when a \code{geolist} is passed to the function, each element of the
\code{geolist} is assumed to be an ensemble member.
}
\examples{
p_ge_0.5 <- ens_prob(ens_grid_df, 0.5)
image(p_ge_0.5$prob_ge_0.5[[1]])

p_le_0.1 <- ens_prob(ens_grid_df, 0.1, comparator = "le")
image(p_le_0.1$prob_le_0.1[[1]])

p_btw_0.25_0.75 <- ens_prob(
  ens_grid_df, c(0.25, 0.75), comparator = "between"
)
image(p_btw_0.25_0.75$prob_between_0.25_0.75[[1]])
}
