% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_df_list.R
\name{ens_prob}
\alias{ens_prob}
\title{Compute the ensemble probability for a threshold}
\usage{
ens_prob(
  x,
  threshold = 0,
  comparator = c("ge", "gt", "le", "lt", "between", "outside"),
  include_low = TRUE,
  include_high = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{harp_ens_grid_df} or \code{harp_ens_point_df} data frame, a \code{geolist},
or a \code{harp_list} containing ensemble data frames.}

\item{threshold}{A threshold for computing binary probabilities. If
\code{comparator = "between"}, it must be a two element vector. Set to NA
to use the raw data.}

\item{comparator}{How to compare \code{x} with the threshold to compute
binary probabilities. Can be \code{"ge"}, \code{"gt"}, \code{"le"}, or
\code{"lt"} for >=, >, <= and < respectively. Can also be
\code{"between"} or \code{"outside"}, in which case the binary probability
of being between or outside of the two values given in \code{threshold} is
computed.}

\item{include_low}{Logical. Whether to include to the lower of the two
thresholds in the comparison when \code{comparator = "between"} or
\code{comparator = "outside"}.}

\item{include_high}{Logical. Whether to include to the higher of the two
thresholds in the comparison when \code{comparator = "between"} or
\code{comparator = "outside"}.}
}
\value{
An object of the same class as \code{x} with the probabilities computed
}
\description{
The probability for a threshold for an ensemble is computed. This is by
default for threshold exceedence (P(fcst >= threshold)), but the probability
for being below the threshold, or between or outside of two thresholds can
also be calculated.
}
\details{
Note that when a \code{geolist} is passed to the function, each element of the
\code{geolist} is assumed to be an ensemble member.
}
\examples{
p_ge_0.5 <- ens_prob(ens_grid_df, 0.5)
image(p_ge_0.5$prob_ge_0.5[[1]])

p_le_0.1 <- ens_prob(ens_grid_df, 0.1, comparator = "le")
image(p_le_0.1$prob_le_0.1[[1]])

p_btw_0.25_0.75 <- ens_prob(
  ens_grid_df, c(0.25, 0.75), comparator = "between"
)
image(p_btw_0.25_0.75$prob_between_0.25_0.75[[1]])
}
