% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\alias{det_point_df}
\alias{ens_point_df}
\alias{det_grid_df}
\alias{ens_grid_df}
\alias{det_point_list}
\alias{ens_point_list}
\alias{det_grid_list}
\alias{ens_grid_list}
\alias{anl_grid_df}
\title{Data for examples}
\format{
A \code{harp_df} data frame or \code{harp_list}
\describe{
\item{fcst_dttm}{Forecast start date-time}
\item{lead_time}{Forecast lead time in hours}
\item{valid_dttm}{Forecast valid date-time}
\item{SID}{ID of observation station}
\item{point_det}{Deterministic point forecast value}
\item{point_mbr000}{Point forecast value for ensemble member 000}
\item{point_mbr001}{Point forecast value for ensemble member 001}
\item{grid_det}{Deterministic gridded forecast}
\item{grid_mbr000}{Gridded forecast for ensemble member 000}
\item{grid_mbr001}{Gridded forecast for ensemble member 001}
\item{grid_anl}{Gridded analysis}
}
}
\usage{
det_point_df

ens_point_df

det_grid_df

ens_grid_df

det_point_list

ens_point_list

det_grid_list

ens_grid_list

anl_grid_df
}
\description{
Data used for demonstrating function usage in examples. The data are random
numbers sampled from a uniform distribution between 0 and 1.
}
\keyword{datasets}
