% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_transform.R
\name{geo_opts}
\alias{geo_opts}
\alias{geo_opts_points}
\alias{geo_opts_regrid}
\alias{geo_opts_subgrid}
\alias{geo_opts_zoom}
\alias{geo_opts_xsection}
\title{Options for different transformations}
\usage{
geo_opts_points(
  points,
  method = c("bilinear", "nearest", "bicubic"),
  mask = NULL,
  force = FALSE,
  weights = NULL,
  keep_weights = FALSE
)

geo_opts_regrid(
  new_grid,
  method = c("bilinear", "nearest", "bicubic"),
  mask = NULL,
  new_mask = NULL,
  weights = NULL,
  keep_weights = FALSE
)

geo_opts_subgrid(i1, i2, j1, j2)

geo_opts_zoom(centre_lon, centre_lat, radius_x, radius_y)

geo_opts_xsection(
  p1,
  p2,
  n = 100,
  method = c("bilinear", "nearest", "bicubic"),
  weights = NULL,
  keep_weights = FALSE
)
}
\arguments{
\item{points}{A data frame of geographic points to which to interpolate the
gridded data. The data frame must include the columns "SID" for a unique id
for the point, "lon" for the longitude of the point in decimal degrees and
"lat" for the latitude of the point in decimal degrees. The data frame can
contain other columns, which will be retained in the output.}

\item{method}{The interpolation method. Can be "nearest" for nearest
neighbour, "bilinear", or "bicubic." The default is "bilinear".}

\item{mask}{A mask to prevent grid points being used in the interpolation.
Should be on the same grid as \code{x} and grid points with values of 0 or FALSE
will be masked from the interpolation.}

\item{force}{For interpolations that include a mask, it is possible that a
point will be surrounded by 4 masked points. In this case the mask will be
ignored and all 4 points used in the interpolation (the default). Set
\code{force = TRUE} to force the mask to be applied and set the interpolated to
\code{NA}.}

\item{weights}{Pre-computed weights for the interpolation. Should be the
output from the appropriate \link{geo_weights} function.}

\item{keep_weights}{Whether to keep weights in the output. If set to TRUE,
the return object will have a "weights" attribute.}

\item{new_grid}{A \code{geofield} or \code{geodomain} on the grid that \code{x} should be
regridded to. \link{define_domain} can be used to define a new \code{geodomain}.}

\item{new_mask}{A \code{geofield} on the same grid as \code{new_grid} with grid points
that should not be interpolated to set to 0 or FALSE.}

\item{i1}{The x index of the western side of the sub domain.}

\item{i2}{The x index of the eastern side of the sub domain.}

\item{j1}{The y index of the southern side of the sub domain.}

\item{j2}{The y index of the northern side of the sub domain.}

\item{centre_lon}{The longitude in decimal degrees of the centre of the
zoomed grid.}

\item{centre_lat}{The latitude in decimal degrees of the centre of the
zoomed grid.}

\item{n}{The number of equally spaced points along the section. The default
is 100.}
}
\description{
When using \link{geo_transform} the transformation options must be passed as
named list appropriate to the transformation. These functions are used to
generate such named lists.
}
