% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{seq_dttm}
\alias{seq_dttm}
\title{Generate a sequence of date-time strings}
\usage{
seq_dttm(start_dttm, end_dttm, by = "1h")
}
\arguments{
\item{start_dttm}{The date-time at the start of the sequence. Must be a
string or numeric of the form YYYYMMDD, YYYYMMDDhh, YYYYMMDDhhmm, or
YYYYMMDDhhmmss.}

\item{end_dttm}{The date-time at the end of the sequence. Must be a string or
numeric of the form YYYYMMDD, YYYYMMDDhh, YYYYMMDDhhmm, or YYYYMMDDhhmmss.}

\item{by}{Increment of the sequence. If numeric, it is considered to be in
hours, otherwise a string with a number followed by a unit. Units can be
"s", for seconds; "m", for minutes; "h", for hours; or "d", for days.}
}
\value{
A sequence of date-time strings
}
\description{
Given a start date-time, end date-time and time resolution a regular sequence
of date-time strings is generated. The start and end date-times must be a
string or numeric of the form YYYYMMDD, YYYYMMDDhh, YYYYMMDDhhmm, or
YYYYMMDDhhmmss.
}
\details{
The output sequence is a vector of strings. Truncation of the strings is done
so that the last zero values are removed.
}
\examples{
seq_dttm(20220306, 20220307)
seq_dttm(20220306, 20220307, by = "30m")
seq_dttm(20220301, 20220331, by = "1d")
seq_dttm(202203061030, 202203061045, by = "30s")
}
